refactor: Update import paths for Algolia components and fix lazy loading for product colors
